+++ src/components/app/App.js
@@ -1,36 +1,21 @@
 import AppHeader from "../appHeader/AppHeader";
-import RandomChar from "../randomChar/RandomChar";
-import CharList from "../charList/CharList";
-import CharInfo from "../charInfo/CharInfo";
-import {useState} from 'react';
-import decoration from '../../resources/img/vision.png';
-import ErrorBoundary from "../errorBoundary/ErrorBoundary";
+import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
+import {ComicsPage, MainPage} from "../../pages";
 
 const App = () => {
-  let [selectedChar, setSelectedChar] = useState(null);
-
-  const onCharSelected = (id) => {
-    setSelectedChar(id);
-  }
 
   return (
-    <div className="app">
-      <AppHeader/>
-      <main>
-        <ErrorBoundary>
-          <RandomChar/>
-        </ErrorBoundary>
-        <div className="char__content">
-          <ErrorBoundary>
-            <CharList onCharSelected={onCharSelected}/>
-          </ErrorBoundary>
-          <ErrorBoundary>
-            <CharInfo charId={selectedChar}/>
-          </ErrorBoundary>
-        </div>
-        <img className="bg-decoration" src={decoration} alt="vision"/>
-      </main>
-    </div>
+    <Router>
+      <div className="app">
+        <AppHeader/>
+        <main>
+          <Routes>
+            <Route path="/" element={<MainPage/>} />
+            <Route path="/comics" element={<ComicsPage/>} />
+          </Routes>
+        </main>
+      </div>
+    </Router>
   )
 }
