+++ src/pages/ComicsPage.js
@@ -1,0 +1,15 @@
+import AppBanner from "../components/appBanner/AppBanner";
+import ComicsList from "../components/comicsList/ComicsList";
+import React from "react";
+
+
+const ComicsPage = () => {
+  return (
+    <>
+      <AppBanner/>
+      <ComicsList/>
+    </>
+  )
+}
+
+export default ComicsPage;
