+++ src/pages/index.js
@@ -1,0 +1,4 @@
+import ComicsPage from "./ComicsPage";
+import MainPage from "./MainPage";
+
+export { MainPage, ComicsPage };
